vector-embedding/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.35.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
redis==5.0.1
numpy==1.24.3
scikit-learn==1.3.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
asyncio-throttle==1.0.2
tenacity==8.2.3
uvloop==0.19.0